gsa: introduce structure to support gsa-based attribution methods
